refactor: 작품 소개 관련 클래스 패키지 분리
