+++ backend/forgather/src/main/java/com/forgather/domain/product/controller/ProductController.java
@@ -1,4 +1,4 @@
-package com.forgather.domain.product;
+package com.forgather.domain.product.controller;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -7,6 +7,10 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+
+import com.forgather.domain.product.service.ProductService;
+import com.forgather.domain.product.dto.CreateProductRequest;
+import com.forgather.domain.product.dto.CreateProductResponse;
 
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
