+++ backend/forgather/src/main/java/com/forgather/domain/product/dto/CreatePhotoRequest.java
@@ -1,4 +1,7 @@
-package com.forgather.domain.product;
+package com.forgather.domain.product.dto;
+
+import com.forgather.domain.product.model.Product;
+import com.forgather.domain.product.model.ProductPhoto;
 
 public record CreatePhotoRequest(
     String originalName,
