+++ backend/forgather/src/main/java/com/forgather/domain/product/dto/CreateProductResponse.java
@@ -1,6 +1,9 @@
-package com.forgather.domain.product;
+package com.forgather.domain.product.dto;
 
 import java.util.List;
+
+import com.forgather.domain.product.model.Product;
+import com.forgather.domain.product.model.ProductPhoto;
 
 public record CreateProductResponse(
     String title,
