+++ backend/forgather/src/main/java/com/forgather/domain/product/repository/ProductPhotoRepository.java
@@ -1,6 +1,6 @@
 package com.forgather.domain.product.repository;
 
-import com.forgather.domain.product.ProductPhoto;
+import com.forgather.domain.product.model.ProductPhoto;
 
 public interface ProductPhotoRepository {
     ProductPhoto save(ProductPhoto productPhoto);
