+++ backend/forgather/src/main/java/com/forgather/domain/product/repository/ProductRepository.java
@@ -1,6 +1,6 @@
 package com.forgather.domain.product.repository;
 
-import com.forgather.domain.product.Product;
+import com.forgather.domain.product.model.Product;
 
 public interface ProductRepository {
     Product save(Product product);
