+++ backend/forgather/src/main/java/com/forgather/domain/product/repository/jpa/ProductJpaRepository.java
@@ -2,7 +2,7 @@ package com.forgather.domain.product.repository.jpa;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.forgather.domain.product.Product;
+import com.forgather.domain.product.model.Product;
 import com.forgather.domain.product.repository.ProductRepository;
 
 public interface ProductJpaRepository extends JpaRepository<Product, Long>, ProductRepository {
